+++ hugo-nginx/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Version bump for new base image
 * Adjust minio-client client.json to config.json
 * Add mirror to S3 script
+* Fix typo in mkdir command
 
 ---